manager updates
